requirements.txt
PySide6>=6.6.0
pyautogui>=0.9.54
pyperclip>=1.8.2
openai>=1.12.0
pyyaml>=6.0
watchdog>=4.0.0
loguru>=0.7.0
pydantic>=2.5.0
pywin32>=306
pillow>=10.0.0
keyboard>=0.13.5

# Optional: voice support
edge-tts>=6.1.0

# Build
PyInstaller>=6.0.0
